backend/requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
